refactor(model): give inverse design its target vocabulary a module
targets.py holds what a search optimises toward and what it hands back: OptimizationTarget, its
validated tensor form, the scalar reduction, and the two result namedtuples. Nothing in it knows
about a model, a simplex or an optimiser.

It is the package's leaf, and that is structural rather than tidy-minded: the two search loops
move to their own modules next, and both need this vocabulary. Leaving it on the mixin that hosts
them would make mixin -> loops -> mixin a cycle. A leaf module is what breaks it.

One duplicate goes with it. Both searches turned the task_targets / class_targets mappings into a
target list with the same twenty lines — optimize_composition's copy is even commented "mirrors
optimize_latent". That is now targets_from_mappings, called by both. What genuinely differs stays
at the call sites: optimize_latent also carries the legacy task_name + mode / target_value path,
so its mutual-exclusion rule names a third keyword, and folding that into a shared helper would
have contorted the helper to serve two policies.
